måtte endre på en del ting, you know what :P

+++ src/main/java/internal/dtos/OrderProductInputDto.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 
 public record OrderProductInputDto(
         long productId,
-        BigDecimal quantity
+        int quantity
 ){
 }
